fix: renamed a file, and changed the custom assert cuntion
