+++ header/fileToString/fileToString.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include <string>
+
+/// <summary>
+/// returns the file into a sting
+/// </summary>
+/// <returns>
+/// returns a string element from the std namespace.
+/// <para>ex: fileToString("./SomeFile.c") returns all it's text in a string element</para>
+/// </returns>
+std::string fileToString(const std::string& filePath);
+
