--- header/io/io.h
@@ -1,8 +1,0 @@
-#pragma once
-#include <string>
-
-namespace io
-{
-	std::string fileToString(const std::string& filePath);
-}
-
